OpenHapticsRenderer: Corrected the deletion of dummy_context. dummy_context is needed to allow for switching between renderers but was incorrectly removed on exit.
Corrected deletion of HAPIHapticShape and HAPISurfaceObject from h3dapi. No longer any problems when touching a surface in h3dapi and exiting. This was done by:
Moving definitions of IntersectionInfo and PlaneConstraint to separate files to allow for AutoRef of HAPIHapticShape in PlaneConstraint.
Made HAPISurfaceObject inherit from RefCountedClass.
Give the possibility to send a clean_up_function to HAPIHapticShape that will be called in destructor.

git-svn-id: https://www.h3d.org:8090/HAPI/trunk@179 f896c5a5-350d-0410-967e-bc877d8f5295