Added HapticTexturesSurface.
Added new function in HAPIFrictionSurface which is intended to replace another function which is deprecated. The reason for doing this is to be able to simply calculate friction surface for a new surface class by only modifying the way the parameters of a hapifrictionsurface are set.

Updated ImageInterfaceObject.cpp. It did not need to be an interface class so I made a function public.

Updated other texture classes because of change in HAPIFrictionSurface.h.

git-svn-id: https://www.h3dapi.org:8090/HAPI/trunk@444 f896c5a5-350d-0410-967e-bc877d8f5295