Patch from bug report 0000232 applied to NiFalconHapticsDevice. To make order of buttons equal for NiFalconHapticsDevice and FalconHapticsDevice.
The rest of the files contains a part fix for bug reports 0000220 and 0000221. Only partly fixed because it fixes the problem with duplicate binary directories by changing a bit in the system for replacing paths in examples and moving a CMakeLists.txt to a separate folder. It does not deal with install order since that is not tested yet. It works fine on my Windows system.

git-svn-id: https://www.h3d.org:8090/HAPI/trunk@549 f896c5a5-350d-0410-967e-bc877d8f5295