Updates to use CPack to divide HAPI into components when packaging. Moved most of the CPack setup to separate file HAPICPack.cmake. Updated NSIS.template.in to the same version as in CMake 2.8.9 since it handles component packages. Change default installation to add to path for current user by default (not all as before).
Updates to FindGLUTWin.cmake to first find freeglut quietly on windows. To not print out a warning about not finding freeglut when it is actually found in external.

Updates to subdirectory adding to add the subdirectory builds in a more logical place than before. H3DUtil cmake build files will be added as a subdirectory to the location of building HAPI. Same kind of updates for examples.

Changed HAPI.cmake to use a cmakedefine for HAVE_OPENGL, that was not done before even though it should have since HAPI does not really require OpenGL.

git-svn-id: https://www.h3dapi.org:8090/HAPI/trunk@623 f896c5a5-350d-0410-967e-bc877d8f5295