Due to a cryptic OpenHaptics bug HLThread is now rewritten to only call hdscheduleasynchronous once and then handle the callback locking for asynchronousCallback, synchronousCallback and removeAsynchronousCallback function in a similar way to the H3DUtil::PeriodicThread. I only copied code from Thread.h and Thread.cpp instead of using some kind of virtual functions. The reason for this is to not have to rearrange PeriodicThread.
There is one really ugly thing I did here which might have to be looked at later. In order to not include HD/hd.h in HLThread.h I simply made the main thread function a public member of HLThread which means that other classes can access it which they really should not. There is a warning in its description though. The only other solution I can think of is to wrap the function in some other friend struct which has public functions to access it, but that will also expose it where it should not be exposed so I choose to just trust programmers to read documentation, hrmph.

git-svn-id: https://www.h3dapi.org:8090/HAPI/trunk@792 f896c5a5-350d-0410-967e-bc877d8f5295