CollisionObjects.cpp - One of the movingSphereIntersect functions was weirdly implemented.. It looks like it is copied from the one which expects a result back when this version only tells if there is intersection or not (and therefore does early exit.
HLThread - Fixed bug in which synchronousCallback did not check if thread was active and therefore a deadlock would be caused (for example if profiling was on). is_active is now only changed within a certain lock. Also added a warning to the isActive function since it may return an incorrect value if called in a separate thread from setActive.

git-svn-id: https://www.h3dapi.org:8090/HAPI/trunk@813 f896c5a5-350d-0410-967e-bc877d8f5295